+++ all/build.gradle
@@ -18,7 +18,7 @@ def subprojects = [
         project(':opencensus-contrib-grpc-util'),
         project(':opencensus-contrib-grpc-metrics'),
         project(':opencensus-contrib-http-util'),
-        project(':opencensus-contrib-log-correlation-log4j'),
+        project(':opencensus-contrib-log-correlation-log4j2'),
         project(':opencensus-contrib-log-correlation-stackdriver'),
         project(':opencensus-contrib-monitored-resource-util'),
         project(':opencensus-contrib-spring'),
@@ -43,7 +43,7 @@ def subprojects_javadoc = [
         project(':opencensus-contrib-grpc-util'),
         project(':opencensus-contrib-grpc-metrics'),
         project(':opencensus-contrib-http-util'),
-        project(':opencensus-contrib-log-correlation-log4j'),
+        project(':opencensus-contrib-log-correlation-log4j2'),
         project(':opencensus-contrib-log-correlation-stackdriver'),
         project(':opencensus-contrib-monitored-resource-util'),
         project(':opencensus-contrib-spring'),
