+++ build.gradle
@@ -154,7 +154,7 @@ subprojects {
         googleAuthVersion = '0.10.0'
         googleCloudBetaVersion = '0.58.0-beta'
         googleCloudGaVersion = '1.40.0'
-        log4jVersion = '2.11.1'
+        log4j2Version = '2.11.1'
         signalfxVersion = '0.0.39'
         springBootVersion = '1.5.15.RELEASE'
         springCloudVersion = '1.3.4.RELEASE'
@@ -177,7 +177,7 @@ subprojects {
                 google_auth: "com.google.auth:google-auth-library-credentials:${googleAuthVersion}",
                 google_cloud_logging: "com.google.cloud:google-cloud-logging:${googleCloudGaVersion}",
                 google_cloud_trace: "com.google.cloud:google-cloud-trace:${googleCloudBetaVersion}",
-                log4j: "org.apache.logging.log4j:log4j-core:${log4jVersion}",
+                log4j2: "org.apache.logging.log4j:log4j-core:${log4j2Version}",
                 zipkin_reporter: "io.zipkin.reporter2:zipkin-reporter:${zipkinReporterVersion}",
                 zipkin_urlconnection: "io.zipkin.reporter2:zipkin-sender-urlconnection:${zipkinReporterVersion}",
                 jaeger_reporter: "com.uber.jaeger:jaeger-core:${jaegerReporterVersion}",
