+++ buildscripts/import-control.xml
@@ -94,7 +94,7 @@ General guidelines on imports:
       <allow pkg="io.opencensus.tags"/>
       <allow pkg="io.opencensus.trace"/>
     </subpackage>
-    <subpackage name="logcorrelation.log4j">
+    <subpackage name="logcorrelation.log4j2">
       <allow pkg="io.opencensus.trace"/>
       <allow pkg="org.apache.logging.log4j"/>
     </subpackage>
