+++ settings.gradle
@@ -19,7 +19,7 @@ include ":opencensus-contrib-exemplar-util"
 include ":opencensus-contrib-grpc-metrics"
 include ":opencensus-contrib-grpc-util"
 include ":opencensus-contrib-http-util"
-include ":opencensus-contrib-log-correlation-log4j"
+include ":opencensus-contrib-log-correlation-log4j2"
 include ":opencensus-contrib-log-correlation-stackdriver"
 include ":opencensus-contrib-monitored-resource-util"
 include ":opencensus-contrib-spring"
@@ -37,8 +37,8 @@ project(':opencensus-contrib-exemplar-util').projectDir = "$rootDir/contrib/exem
 project(':opencensus-contrib-grpc-metrics').projectDir = "$rootDir/contrib/grpc_metrics" as File
 project(':opencensus-contrib-grpc-util').projectDir = "$rootDir/contrib/grpc_util" as File
 project(':opencensus-contrib-http-util').projectDir = "$rootDir/contrib/http_util" as File
-project(':opencensus-contrib-log-correlation-log4j').projectDir =
-        "$rootDir/contrib/log_correlation/log4j" as File
+project(':opencensus-contrib-log-correlation-log4j2').projectDir =
+        "$rootDir/contrib/log_correlation/log4j2" as File
 project(':opencensus-contrib-log-correlation-stackdriver').projectDir =
         "$rootDir/contrib/log_correlation/stackdriver" as File
 project(':opencensus-contrib-monitored-resource-util').projectDir =
